+++ oap-server/analyzer/log-analyzer/src/main/java/org/apache/skywalking/oap/log/analyzer/dsl/spec/AbstractSpec.java
@@ -22,6 +22,7 @@ import groovy.lang.Closure;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.Accessors;
+import org.apache.skywalking.apm.network.common.v3.KeyStringValuePair;
 import org.apache.skywalking.oap.log.analyzer.dsl.Binding;
 import org.apache.skywalking.oap.log.analyzer.provider.LogAnalyzerModuleConfig;
 import org.apache.skywalking.oap.server.library.module.ModuleManager;
@@ -48,5 +49,15 @@ public abstract class AbstractSpec {
     @SuppressWarnings("unused")
     public Object propertyMissing(final String name) {
         return BINDING.get().getVariable(name);
+    }
+
+    @SuppressWarnings("unused")
+    public String tag(String key) {
+        return BINDING.get().log().getTags().getDataList()
+                .stream()
+                .filter(data -> key.equals(data.getKey()))
+                .map(KeyStringValuePair::getValue)
+                .findFirst()
+                .orElse("");
     }
 }